% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmalloc_ld.R
\name{ld_ncol}
\alias{ld_ncol}
\title{Number of variants in a banded LD store}
\usage{
ld_ncol(store)
}
\arguments{
\item{store}{An \link{fmalloc_ld} object.}
}
\value{
The number of variants (columns == rows) as an integer.
}
\description{
Number of variants in a banded LD store
}
\seealso{
\code{\link[=fmalloc_ld]{fmalloc_ld()}}, \code{\link[=ld_pair]{ld_pair()}}, \code{\link[=ld_col]{ld_col()}}
}
