% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmalloc.R
\name{cleanup_fmalloc}
\alias{cleanup_fmalloc}
\title{Clean Up fmalloc}
\usage{
cleanup_fmalloc()
}
\value{
NULL (invisibly)
}
\description{
Cleans up the fmalloc allocator and releases resources.
Call this when you're done using fmalloc-based allocation.
}
\examples{
\dontrun{
init_fmalloc("data.bin")
v <- create_fmalloc_vector("integer", 100)
# ... use vector ...
cleanup_fmalloc()
}

}
